CascadeProjects/windsurf-project/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
anthropic==0.8.1
openai==1.3.7
google-generativeai==0.3.2
groq==0.4.1
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0
httpx==0.25.2
